web-animations/animation-player/animation-player-current-time.html

<!DOCTYPE html>
<meta charset=utf-8>
<title>Animation player currentTime attribute test</title>
<meta name="assert" content="The effective current time of this player. Setting this attribute follows the procedure defined in section 3.5.2.9 Performing a seek">
<link rel="help" href="http://dev.w3.org/fxtf/web-animations/#widl-AnimationPlayer-currentTime">
<link rel="author" title="Sergey G. Grekhov" href="mailto:sgrekhov@unipro.ru">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../testcommon.js"></script>
<link rel="stylesheet" href="/resources/testharness.css">
<body>
<div id="log"></div>
<iframe id="iframe" src="../resources/blank.html"></iframe>
<script>
var CONTENT_END_TIME = 1000;

function newAnimation(doc) {
    var animationTarget = doc.createElement('div');
    animationTarget.style.width = '300px';
    doc.body.appendChild(animationTarget);
    return new Animation(animationTarget, [
	    {width: "10px", offset: 0},
	    {width: "100px", offset: 1/2},
	    {width: "200px", offset: 1}], CONTENT_END_TIME);
}


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));

    assert_not_equals(player.currentTime, null,
        'The value of AnimationPlayer.currentTime attribute should not be of type null');
}, 'Test value of AnimationPlayer.currentTime attribute should not be null');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));

    assert_throws('TypeError',function () {
        player.currentTime = [];
    }, 'TypeError should be thrown if AnimationPlayer.currentTime attribute ' +
        'is set to not double value');
}, 'Test that TypeError is thrown if AnimationPlayer.currentTime attribute ' +
    'is set to not double value');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));

    var newCurrentTime = player.currentTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));
    player.pause();

    var newCurrentTime = player.currentTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute. Player is paused');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));
    player.playbackRate = 0;

    var newCurrentTime = player.currentTime + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute. ' +
    'Player playback rate is zero');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));
    player.playbackRate = -1;

    var newCurrentTime = -100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute. ' +
    'Player playback rate is < 0 and seek time < 0');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));

    var newCurrentTime = CONTENT_END_TIME;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute. ' +
    'Player playback rate is > 0 and seek time equals to the source content end');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));

    var newCurrentTime = CONTENT_END_TIME + 100;
    player.currentTime = newCurrentTime;

    assert_equals(player.currentTime, newCurrentTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
}, 'Test setting value of AnimationPlayer.currentTime attribute. ' +
    'Player playback rate is > 0 and seek time > source content end');


test(function() {
    var doc = newHTMLDocument();
    var player = doc.timeline.play(newAnimation(doc));

    var seekTime = player.currentTime + 100;
    var expectedTime = player.timeline.currentTime - seekTime / player.playbackRate;
    player.currentTime = seekTime;

    assert_equals(player.currentTime, expectedTime,
        'The value of AnimationPlayer.currentTime attribute should be set to the new value');
    assert_false(player.finished, 'Player\s finished flag should be set to false');
}, 'Test setting value of AnimationPlayer.currentTime attribute. ' +
    'Player is not paused and playback rate is not zero');


test(function() {
    var iframe = document.querySelector('#iframe');
    var doc = iframe.contentWindow.document;
    var animation = newAnimation(doc);
    var player = doc.timeline.play(animation);
    player.currentTime = 0;

    assert_equals(window.getComputedStyle(animation.target).width, 10,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute. Current time is zero');


test(function() {
    var iframe = document.querySelector('#iframe');
    var doc = iframe.contentWindow.document;
    var animation = newAnimation(doc);
    var player = doc.timeline.play(animation);
    player.currentTime = 500;

    assert_equals(window.getComputedStyle(animation.target).width, 100,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute. Current time is > 0 and < end time');


test(function() {
    var iframe = document.querySelector('#iframe');
    var doc = iframe.contentWindow.document;
    var animation = newAnimation(doc);
    var player = doc.timeline.play(animation);
    player.currentTime = CONTENT_END_TIME;

    assert_equals(window.getComputedStyle(animation.target).width, 200,
        'Animation effect should take place');
}, 'Test that animation effect is changed accordingly by setting ' +
    'AnimationPlayer.currentTime attribute. Current time is > 0 and < end time');

document.querySelector('#iframe').remove();
</script>
</body>
